+++ apps/assets/src/components/asset/modals/transactions/hooks/useDeposit.ts
@@ -28,11 +28,27 @@ import {
   EVMOS_SYMBOL,
   StoreType,
 } from "evmos-wallet";
+import {
+  CLICK_DEPOSIT_CONFIRM_BUTTON,
+  useTracker,
+  SUCCESSFUL_TX_DEPOSIT,
+  UNSUCCESSFUL_TX_DEPOSIT,
+} from "tracker";
 export const useDeposit = (useDepositProps: DepositProps) => {
   const wallet = useSelector((state: StoreType) => state.wallet.value);
   const dispatch = useDispatch();
-
+  const { handlePreClickAction } = useTracker(CLICK_DEPOSIT_CONFIRM_BUTTON);
+  const { handlePreClickAction: successfulTx } = useTracker(
+    SUCCESSFUL_TX_DEPOSIT
+  );
+  const { handlePreClickAction: unsuccessfulTx } = useTracker(
+    UNSUCCESSFUL_TX_DEPOSIT
+  );
   const handleConfirmButton = async () => {
+    handlePreClickAction({
+      wallet: wallet?.evmosAddressEthFormat,
+      provider: wallet?.extensionName,
+    });
     useDepositProps.setConfirmClicked(true);
     if (wallet.osmosisPubkey === null) {
       dispatch(snackRequestRejected());
@@ -126,6 +142,14 @@ export const useDeposit = (useDepositProps: DepositProps) => {
     );
 
     dispatch(snackExecuteIBCTransfer(res));
+    if (res.error) {
+      unsuccessfulTx({
+        errorMessage: res.message,
+        wallet: wallet?.evmosAddressEthFormat,
+        provider: wallet?.extensionName,
+        transaction: "unsuccessful",
+      });
+    }
     useDepositProps.setShow(false);
     // check if tx is executed
     if (res.title === BROADCASTED_NOTIFICATIONS.SuccessTitle) {
@@ -144,6 +168,12 @@ export const useDeposit = (useDepositProps: DepositProps) => {
           chainIds.chainIdentifier.toUpperCase()
         )
       );
+      successfulTx({
+        txHash: res.txHash,
+        wallet: wallet?.evmosAddressEthFormat,
+        provider: wallet?.extensionName,
+        transaction: "successful",
+      });
     }
   };