+++ apps/staking/src/components/staking/modals/hooks/useRedelegation.tsx
@@ -7,12 +7,27 @@ import { parseUnits } from "@ethersproject/units";
 import { BigNumber } from "ethers";
 import { executeRedelegate } from "../../../../internal/staking/functionality/transactions/redelegate";
 import { snackExecuteIBCTransfer } from "evmos-wallet";
-
+import {
+  CLICK_BUTTON_CONFIRM_REDELEGATE,
+  useTracker,
+  SUCCESSFUL_TX_REDELEGATE,
+  UNSUCCESSFUL_TX_REDELEGATE,
+} from "tracker";
 export const useRedelegation = (useRedelegateProps: RedelegateProps) => {
   const dispatch = useDispatch();
 
-  //   async
+  const { handlePreClickAction } = useTracker(CLICK_BUTTON_CONFIRM_REDELEGATE);
+  const { handlePreClickAction: successfulTx } = useTracker(
+    SUCCESSFUL_TX_REDELEGATE
+  );
+  const { handlePreClickAction: unsuccessfulTx } = useTracker(
+    UNSUCCESSFUL_TX_REDELEGATE
+  );
   const handleConfirmButton = async () => {
+    handlePreClickAction({
+      wallet: useRedelegateProps?.wallet?.evmosAddressEthFormat,
+      provider: useRedelegateProps?.wallet?.extensionName,
+    });
     useRedelegateProps.setConfirmClicked(true);
     if (
       useRedelegateProps.value === undefined ||
@@ -39,6 +54,21 @@ export const useRedelegation = (useRedelegateProps: RedelegateProps) => {
       useRedelegateProps.validatorDst
     );
     dispatch(snackExecuteIBCTransfer(res));
+    if (res.error === true) {
+      unsuccessfulTx({
+        errorMessage: res.message,
+        wallet: useRedelegateProps.wallet?.evmosAddressEthFormat,
+        provider: useRedelegateProps.wallet?.extensionName,
+        transaction: "unsuccessful",
+      });
+    } else {
+      successfulTx({
+        txHash: res.txHash,
+        wallet: useRedelegateProps.wallet?.evmosAddressEthFormat,
+        provider: useRedelegateProps.wallet?.extensionName,
+        transaction: "successful",
+      });
+    }
     useRedelegateProps.setShow(false);
   };