+++ apps/staking/src/components/staking/modals/hooks/useUndelegations.tsx
@@ -8,10 +8,26 @@ import { BigNumber } from "ethers";
 import { executeUndelegate } from "../../../../internal/staking/functionality/transactions/undelegate";
 import { snackExecuteIBCTransfer } from "evmos-wallet";
 
+import {
+  CLICK_BUTTON_CONFIRM_UNDELEGATE,
+  useTracker,
+  SUCCESSFUL_TX_UNDELEGATE,
+  UNSUCCESSFUL_TX_UNDELEGATE,
+} from "tracker";
 export const useUndelegation = (useUndelegateProps: UndelegateProps) => {
   const dispatch = useDispatch();
-
+  const { handlePreClickAction } = useTracker(CLICK_BUTTON_CONFIRM_UNDELEGATE);
+  const { handlePreClickAction: successfulTx } = useTracker(
+    SUCCESSFUL_TX_UNDELEGATE
+  );
+  const { handlePreClickAction: unsuccessfulTx } = useTracker(
+    UNSUCCESSFUL_TX_UNDELEGATE
+  );
   const handleConfirmButton = async () => {
+    handlePreClickAction({
+      wallet: useUndelegateProps?.wallet?.evmosAddressEthFormat,
+      provider: useUndelegateProps?.wallet?.extensionName,
+    });
     useUndelegateProps.setConfirmClicked(true);
     if (
       useUndelegateProps.value === undefined ||
@@ -35,6 +51,21 @@ export const useUndelegation = (useUndelegateProps: UndelegateProps) => {
       amount
     );
     dispatch(snackExecuteIBCTransfer(res));
+    if (res.error === true) {
+      unsuccessfulTx({
+        errorMessage: res.message,
+        wallet: useUndelegateProps.wallet?.evmosAddressEthFormat,
+        provider: useUndelegateProps.wallet?.extensionName,
+        transaction: "unsuccessful",
+      });
+    } else {
+      successfulTx({
+        txHash: res.txHash,
+        wallet: useUndelegateProps.wallet?.evmosAddressEthFormat,
+        provider: useUndelegateProps.wallet?.extensionName,
+        transaction: "successful",
+      });
+    }
     useUndelegateProps.setShow(false);
   };