+++ packages/evmos-wallet/src/wallet/ButtonWalletConnection.tsx
@@ -19,7 +19,12 @@ import { Keplr } from "../internal/wallet/functionality/keplr/keplr";
 import { disconnectWallets } from "../internal/wallet/functionality/disconnect";
 import {
   GetProviderFromLocalStorage,
+  GetProviderWalletConnectFromLocalStorage,
+  GetWalletFromLocalStorage,
   RemoveProviderFromLocalStorage,
+  RemoveProviderWalletConnectToLocalStorage,
+  RemoveWalletFromLocalStorage,
+  SaveWalletToLocalStorage,
 } from "../internal/wallet/functionality/localstorage";
 import {
   KeplrIcon,
@@ -37,7 +42,15 @@ import {
   useWalletConnect,
 } from "../internal/wallet/functionality/walletconnect/walletconnect";
 import { Tooltip } from "ui-helpers";
-
+import {
+  CLICK_WC_CONNECT_WALLET_BUTTON,
+  CLICK_WC_DISCONNECT_WALLET_BUTTON,
+  CLICK_WC_CONNECTED_WITH,
+  SWITCH_BETWEEN_WALLETS,
+  SUCCESSFUL_WALLET_CONNECTION,
+  UNSUCCESSFUL_WALLET_CONNECTION,
+  useTracker,
+} from "tracker";
 // Components
 import { Button } from "ui-helpers";
 
@@ -46,7 +59,7 @@ export const ButtonWalletConnection = ({
   dispatch,
 }: {
   walletExtension: WalletExtension;
-  dispatch: Dispatch<AnyAction>;
+  dispatch: Dispatch<AnyAction>; // eslint-disable-next-line sonarjs/cognitive-complexity
 }) => {
   const [show, setShow] = useState(false);
 
@@ -91,6 +104,69 @@ export const ButtonWalletConnection = ({
   });
 
   const [isCopied, setIsCopied] = useState(false);
+
+  const { handlePreClickAction: trackClickConnectWallet } = useTracker(
+    CLICK_WC_CONNECT_WALLET_BUTTON
+  );
+  const { handlePreClickAction: trackClickDisconnectWallet } = useTracker(
+    CLICK_WC_DISCONNECT_WALLET_BUTTON
+  );
+
+  const { handlePreClickAction: trackConnectedWithWallet } = useTracker(
+    CLICK_WC_CONNECTED_WITH
+  );
+
+  const { handlePreClickAction: trackChangeWallet } = useTracker(
+    SWITCH_BETWEEN_WALLETS
+  );
+
+  const { handlePreClickAction: trackSuccessfulWalletConnection } = useTracker(
+    SUCCESSFUL_WALLET_CONNECTION
+  );
+
+  const { handlePreClickAction: trackUnsuccessfulWalletConnection } =
+    useTracker(UNSUCCESSFUL_WALLET_CONNECTION);
+  useEffect(() => {
+    function trackWallet() {
+      const walletLocalStorage = GetWalletFromLocalStorage();
+      // walletExtension is not set
+      if (walletExtension.evmosAddressEthFormat === "") {
+        return;
+      }
+      // walletLocalStorage is not set
+      if (walletLocalStorage === null) {
+        return;
+      }
+      // track the wallet change if the wallets are different
+      if (walletExtension.evmosAddressEthFormat !== walletLocalStorage) {
+        trackChangeWallet({
+          provider: walletExtension.extensionName,
+          wallet: walletExtension.evmosAddressEthFormat,
+        });
+        SaveWalletToLocalStorage(walletExtension.evmosAddressEthFormat);
+      }
+    }
+    // tracking address changes
+    if (METAMASK_KEY === GetProviderFromLocalStorage()) {
+      trackWallet();
+    }
+
+    if (KEPLR_KEY === GetProviderFromLocalStorage()) {
+      trackWallet();
+    }
+  }, [walletExtension]);
+
+  useEffect(() => {
+    const walletLocalStorage = GetWalletFromLocalStorage();
+    // avoid saving the evmos address if it is empty or is already stored.
+    if (walletExtension.evmosAddressEthFormat === "") {
+      return;
+    }
+    if (walletLocalStorage === walletExtension.evmosAddressEthFormat) {
+      return;
+    }
+    SaveWalletToLocalStorage(walletExtension.evmosAddressEthFormat);
+  }, [walletExtension]);
 
   return walletExtension.active === true ? (
     <>
@@ -164,7 +240,13 @@ export const ButtonWalletConnection = ({
             <button
               className="w-full rounded font-bold uppercase border border-darkPearl hover:bg-grayOpacity p-3 mt-3"
               onClick={() => {
+                trackClickDisconnectWallet({
+                  wallet: walletExtension?.evmosAddressEthFormat,
+                  provider: walletExtension?.extensionName,
+                });
+                RemoveWalletFromLocalStorage();
                 RemoveProviderFromLocalStorage();
+                RemoveProviderWalletConnectToLocalStorage();
                 disconnectWallets(dispatch);
                 setShow(false);
                 setIsCopied(false);
@@ -178,7 +260,12 @@ export const ButtonWalletConnection = ({
     </>
   ) : (
     <div className="flex justify-center">
-      <Button onClick={open}>
+      <Button
+        onClick={() => {
+          setShow(true);
+          trackClickConnectWallet();
+        }}
+      >
         <div className="flex items-center space-x-2 ">
           <WalletIcon />
           <span>Connect wallet</span>
@@ -195,7 +282,21 @@ export const ButtonWalletConnection = ({
                 setShow(false);
                 disconnectWallets(dispatch);
                 const keplr = new Keplr(store);
-                await keplr.connect();
+                const resultConnect = await keplr.connect();
+                trackConnectedWithWallet({
+                  wallet: GetWalletFromLocalStorage(),
+                  provider: KEPLR_KEY,
+                });
+                if (resultConnect.result) {
+                  trackSuccessfulWalletConnection({
+                    provider: KEPLR_KEY,
+                  });
+                } else {
+                  trackUnsuccessfulWalletConnection({
+                    message: resultConnect.message,
+                    provider: KEPLR_KEY,
+                  });
+                }
               }}
             >
               <ContentModalConnect>
@@ -209,7 +310,21 @@ export const ButtonWalletConnection = ({
                 setShow(false);
                 disconnectWallets(dispatch);
                 const metamask = new Metamask(store);
-                await metamask.connect();
+                const resultConnect = await metamask.connect();
+                trackConnectedWithWallet({
+                  wallet: GetWalletFromLocalStorage(),
+                  provider: METAMASK_KEY,
+                });
+                if (resultConnect.result) {
+                  trackSuccessfulWalletConnection({
+                    provider: METAMASK_KEY,
+                  });
+                } else {
+                  trackUnsuccessfulWalletConnection({
+                    message: resultConnect.message,
+                    provider: METAMASK_KEY,
+                  });
+                }
               }}
             >
               <ContentModalConnect>
@@ -222,6 +337,11 @@ export const ButtonWalletConnection = ({
               onClick={async () => {
                 setShow(false);
                 await useWC.connect();
+                trackConnectedWithWallet({
+                  wallet: GetWalletFromLocalStorage(),
+                  provider: WALLECT_CONNECT_KEY,
+                  walletSelected: GetProviderWalletConnectFromLocalStorage(),
+                });
               }}
             >
               <ContentModalConnect>