+++ packages/tracker/src/constants.ts
@@ -1,0 +1,112 @@
+// mission control
+
+// add the action on the name and in the description
+export const CLICK_REDIRECTS_MISSION_CONTROL_SIDEBAR =
+  "Go on a dApp page from Mission Control";
+export const CLICK_MISSION_CONTROL_HALF_LIFE =
+  "Click on Half-Life from Mission Control";
+export const CLICK_MISSION_CONTROL_VIEW_ASSETS_BUTTON =
+  "Click on View All Assets from Mission Control";
+export const CLICK_MISSION_CONTROL_GOVERNANCE_VOTE_BUTTON =
+  "Click on Vote from Mission Control";
+export const CLICK_MISSION_CONTROL_GOVERNANCE_DOCS_BUTTON =
+  "Click on Docs from Mission Control";
+export const CLICK_MISSION_CONTROL_GOVERNANCE_DISCUSSION_BUTTON =
+  "Click on Discussion from Mission Control";
+export const CLICK_MISSION_CONTROL_ECOSYSTEM_BUTTON =
+  "Click on Ecosystem from Mission Control";
+export const LISTEN_MISSION_CONTROL_SCROLL_ECOSYSTEM =
+  "Scroll through Ecosystem from Mission Control";
+export const CLICK_MISSION_CONTROL_VIEW_STAKING_BUTTON =
+  "Click on View Staking from Mission Control";
+export const CLICK_MISSION_CONTROL_TOTAL_STAKED_TOPBAR =
+  "Click on Total Staked from Mission Control”";
+
+// generics
+export const CLICK_WC_CONNECT_WALLET_BUTTON = "Connect wallet";
+export const CLICK_WC_DISCONNECT_WALLET_BUTTON = "Disconnect wallet";
+export const CLICK_WC_CONNECTED_WITH = "Connect Wallet with a specific Wallet";
+export const CLICK_BACK_TO_MC = "Click “Back to Mission Control”";
+export const CLICK_EVMOS_LOGO = "Click on Evmos Logo";
+export const CLICK_FEEDBACK_FOOTER = "Click on Give Feedback on Footer";
+export const SWITCH_BETWEEN_WALLETS = "Switch between wallets";
+
+// assets
+export const CLICK_DEPOSIT_CHOOSE_FROM_CHAIN = "Deposit from a specific chain";
+export const CLICK_DEPOSIT_CONFIRM_BUTTON = "Confirm deposit transaction";
+export const CLICK_WITHDRAW_CHOOSE_TOKEN = "Choose a token to withdraw";
+export const CLICK_WITHDRAW_CONFIRM_BUTTON = "Confirm withdraw transaction";
+export const CLICK_HIDE_ZERO_BALANCE = "Hide Zero Balance";
+export const CLICK_CTA_LINKS_REGISTER_TOKEN =
+  "Click on CTA link to register token";
+export const CLICK_CTA_LINKS_ASSETS_GUIDE =
+  "Click on CTA link to view assets guide";
+
+// staking
+export const CLICK_CLAIM_REWARDS_TOPBAR = "Claim Evmos Rewards";
+export const CLICK_SHOW_INACTIVE_TOGGLE = "Show inactive";
+export const CLICK_SEARCH_VALIDATORS_INPUT = "Search for a validator";
+export const CLICK_TABS_STAKING_OPTIONS = "Click on a tab on the staking menu";
+export const CLICK_BUTTON_MANAGE_DELEGATE = "Choose to delegate to a validator";
+export const CLICK_BUTTON_MANAGE_UNDELEGATE =
+  "Choose to undelegate to a validator";
+export const CLICK_BUTTON_MANAGE_REDELEGATE =
+  "Choose to redelegate to a validator";
+export const CLICK_BUTTON_CONFIRM_DELEGATE = "Confirm Delegate transaction";
+export const CLICK_BUTTON_CONFIRM_UNDELEGATE = "Confirm Undelegate transaction";
+export const CLICK_BUTTON_CONFIRM_REDELEGATE = "Confirm Redelegate transaction";
+export const CLICK_SORT_VALIDATORS = "Sort through Validators";
+export const CLICK_CANCEL_UNDELEGATION_BUTTON = "Choose to cancel undelegation";
+export const CLICK_CONFIRM_CANCEL_UNDELEGATION_BUTTON =
+  "Confirm cancel undelegation";
+
+// governance
+export const CLICK_BACK_TO_GOVERNANCE = "Back to Governance";
+export const CLICK_COMMONWEALTH_OUTLINK = "Click on Commonwealth outlink";
+export const CLICK_GOVERNANCE_PROPOSAL = "Click on a governance proposal";
+export const CLICK_VOTE_BUTTON = "Choose to vote for a governance proposal";
+export const CLICK_CONFIRM_VOTE_BUTTON = "Confirm vote transaction";
+
+// track un/successful transactions
+export const SUCCESSFUL_TX_DEPOSIT = "Successful transaction of Deposit";
+export const UNSUCCESSFUL_TX_DEPOSIT = "Unsuccessful transaction of Deposit";
+export const SUCCESSFUL_TX_WITHDRAW = "Successful transaction of Withdraw";
+export const UNSUCCESSFUL_TX_WITHDRAW = "Unsuccessful transaction of Withdraw";
+export const SUCCESSFUL_TX_VOTE = "Successful transaction of Vote";
+export const UNSUCCESSFUL_TX_VOTE = "Unsuccessful transaction of Vote";
+
+export const SUCCESSFUL_TX_DELEGATE = "Successful transaction of Delegate";
+export const UNSUCCESSFUL_TX_DELEGATE = "Unsuccessful transaction of Delegate";
+export const SUCCESSFUL_TX_UNDELEGATE = "Successful transaction of Undelegate";
+export const UNSUCCESSFUL_TX_UNDELEGATE =
+  "Unsuccessful transaction of Undelegate";
+export const SUCCESSFUL_TX_REDELEGATE = "Successful transaction of Redelegate";
+export const UNSUCCESSFUL_TX_REDELEGATE =
+  "Unsuccessful transaction of Redelegate";
+export const SUCCESSFUL_TX_CANCEL_UNDELEGATION =
+  "Successful transaction of Cancel Undelegation";
+export const UNSUCCESSFUL_TX_CANCEL_UNDELEGATION =
+  "Unsuccessful transaction of Cancel Undelegation";
+export const SUCCESSFUL_TX_CLAIM_REWARDS =
+  "Successful transaction of Claim Rewards";
+export const UNSUCCESSFUL_TX_CLAIM_REWARDS =
+  "Unsuccessful transaction of Claim Rewards";
+
+export const DISABLE_TRACKER_LOCALSTORAGE = "disableTracker";
+export const DISABLE_MIXPANEL_TRACKER = "Disable mixpanel tracker";
+export const ENABLE_MIXPANEL_TRACKER = "Enable mixpanel tracker";
+
+export const SUCCESSFUL_WALLET_CONNECTION = "Successful wallet connection";
+export const UNSUCCESSFUL_WALLET_CONNECTION = "Unsuccessful wallet connection";
+
+export const CLICK_BUTTON_CONVERT_AXELAR_BASED = "Convert Axelar-based token";
+export const CLICK_BUTTON_CONFIRM_CONVERT_AXELAR_BASED_TX =
+  "Confirm Convert Transaction for Axelar-based token";
+export const SUCCESSFUL_CONVERT_AXELAR_BASED_TX =
+  "Successful convert for Axelar-based token";
+export const UNSUCCESSFUL_CONVERT_AXELAR_BASED_TX =
+  "Unsuccessful convert for Axelar-based token";
+
+export const CLICK_BUTTON_CONFIRM_WRAP_TX = "Confirm wrap transaction";
+export const SUCCESSFUL_WRAP_TX = "Successful wrap transaction";
+export const UNSUCCESSFUL_WRAP_TX = "Unsuccessful wrap transaction";