+++ packages/ui-helpers/src/Footer.tsx
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/apps/blob/main/LICENSE)
 
 import Link from "next/link";
-const VERSION_TAG = "c549d8f";
 const COMMONWEALTH_URL = "https://commonwealth.im/evmos";
 import {
   TwitterIcon,
@@ -11,8 +10,25 @@ import {
   DiscordIcon,
   CommonWealthIcon,
 } from "icons";
+import { PRIVACY_POLICY_URL } from "constants-helper";
 
-export const Footer = () => {
+export const Footer = ({
+  onClickFeedback,
+  handleCookies,
+  version,
+}: {
+  onClickFeedback?: React.MouseEventHandler<HTMLAnchorElement>;
+  handleCookies?: React.MouseEventHandler<HTMLButtonElement>;
+  version: string;
+}) => {
+  function getVersion() {
+    const app_env = process.env.NEXT_PUBLIC_EVMOS_APP_ENV ?? "staging";
+    if (app_env === "production") {
+      return version;
+    }
+    return "main - " + version;
+  }
+
   return (
     <footer className=" mb-10 mt-10 flex w-full flex-col items-center space-y-2 text-pearl xl:justify-between">
       <div className="flex items-center space-x-5">
@@ -58,7 +74,7 @@ export const Footer = () => {
         </Link>
       </div>
       <div className="flex w-full items-center justify-center space-x-5 px-2">
-        <p>Version: {VERSION_TAG}</p>
+        <p>Version: {getVersion()}</p>
         <p>
           <Link
             target="_blank"
@@ -73,7 +89,7 @@ export const Footer = () => {
           <Link
             target="_blank"
             rel="noreferrer"
-            href="https://evmos.org/privacy-policy"
+            href={PRIVACY_POLICY_URL}
             aria-label="privacy policy"
           >
             Privacy Policy
@@ -85,10 +101,12 @@ export const Footer = () => {
             rel="noreferrer"
             href="https://evmos.canny.io/feedback"
             aria-label="feedback"
+            onClick={onClickFeedback}
           >
             Feedback
           </Link>
         </p>
+        <button onClick={handleCookies}>Cookies Settings</button>
       </div>
     </footer>
   );