+++ CHANGELOG.md
@@ -34,7 +34,38 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
-## UNRELEASED
+## 1.0.5 - 2023-05-24
+
+- (workflow) #fse-511 | packages/tracker 1.0.1 | Adding lint
+- (workflow) #fse-511 | packages/constants-helper 1.0.1 packages/eslint-config-custom 1.0.1 packages/evmos-wallet 1.0.5 packages/helpers 1.0.3 packages/icons 1.0.1 packages/services 1.0.2 packages/ui-helpers 1.0.2 | Deleting eslint related dependencies, setting new linting rules, fixing linting issues
+- (workflow) #fse-511 | apps/assets 1.0.3 apps/governance 1.0.3 apps/staking 1.0.3 apps/mission 1.0.3 | Moving husky to the mono repo level
+- (workflow) #fse-511 | husky, package | Updating husky so it can work again
+- (workflow) #fse-511 | apps/mission 1.0.3 | Ignoring \_redirects from git
+- (workflow) #fse-531 | script | Adding script to clear cache and fixing cache issue in the build process
+- (refactor) #fse-514 | github actions | Removing unused folder
+- (chore) #fse-551 | apps/assets 1.0.4 apps/governance 1.0.4 apps/staking 1.0.4 apps/mission 1.0.4 packages/ui-helpers 1.0.3 | add correct env and app version in footer
+- (chore) #fse-581 | apps/assets 1.0.4 | add images for new assets & group quicksilver assets
+- (chore) #fse-519 | apps/assets 1.0.5 apps/governance 1.0.4 apps/staking 1.0.4 apps/mission 1.0.4 | Set up a consent modal for users to opt-in/opt-out to tracking
+- (chore) #fse-519 | packages/evmos-wallet 1.0.7 | Get, set and remove WALLET_KEY and use it for wallet connection tracking
+- (chore) #fse-519 | packages/tracker 1.0.0 | Create tracker package
+- (chore) #fse-519 | packages/ui-helpers 1.0.3 | Update jest configuration. Add onClick function to the components in order to track the events.
+- (workflow) #fse-519 | turbo.json | Add tracker package and NEXT_PUBLIC_MIXPANEL_TOKEN variable
+- (workflow) #fse-519 | package.json 1.0.4 | Add --force --no-cache to build script
+- (chore) #fse-522 | apps/assets 1.0.6 apps/governance 1.0.5 apps/mission 1.0.5 apps/staking 1.0.5 | Add consent modal
+- (chore) #fse-522 | packages/ui-helpers 1.0.4 | Add new constants for privacy policy, cookies policy and TOS version. Add consent modal for TOS.
+- (chore) #fse-522 | packages/tracker 1.0.1 | Enable/disable tracking if DISABLE_TRACKER_LOCALSTORAGE is set
+- (chore) #fse-521 | apps/assets 1.0.7 apps/governance 1.0.6 apps/mission 1.0.6 apps/staking 1.0.6 | Reuse consent modal on footer to allow users to change their opt-in/opt-out preferences
+- (chore) #fse-521 | packages/ui-helpers 1.0.5 | Add Cookies settings button so the user can change their cookies preferences
+- (workflow) #fse-521 | turbo.json | Add NEXT_PUBLIC_MIXPANEL_TOKEN, NEXT_PUBLIC_COOKIE_POLICY_ID_IUBENDA, NEXT_PUBLIC_SITE_ID_IUBENDA
+
+## 1.0.4 - 2023-05-19
+
+- (chore) #fse-581 | apps/assets 1.0.4 | add images for new assets & group quicksilver assets
+- (workflow) #fse-531 | script | Adding script to clear cache and fixing cache issue in the build process
+
+## 1.0.3 - 2023-05-10
+
+- (chore) #fse-142 | evmos-wallet 1.0.6 | Env vars prefixes
 
 - (workflow) #fse-511 | packages/tracker 1.0.1 | Adding lint
 - (workflow) #fse-511 | packages/constants-helper 1.0.1 packages/eslint-config-custom 1.0.1 packages/evmos-wallet 1.0.5 packages/helpers 1.0.3 packages/icons 1.0.1 packages/services 1.0.2 packages/ui-helpers 1.0.2 | Deleting eslint related dependencies, setting new linting rules, fixing linting issues
@@ -42,15 +73,6 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - (workflow) #fse-511 | husky, package | Updating husky so it can work again
 - (workflow) #fse-511 | apps/mission 1.0.3 | Ignoring \_redirects from git
 - (refactor) #fse-514 | github actions | Removing unused folder
-
-
-## 1.0.4 - 2023-05-19
-- (chore) #fse-581 | apps/assets 1.0.4 | add images for new assets & group quicksilver assets
-- (workflow) #fse-531 | script | Adding script to clear cache and fixing cache issue in the build process
-
-## 1.0.3 - 2023-05-10
-- (chore) #fse-142 | evmos-wallet 1.0.6 | Env vars prefixes
-
 
 ## 1.0.2 - 2023-05-10
 
@@ -61,7 +83,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## 1.0.1 - 2023-05-09
 
-- (workflow) #fse-512 | githug/workflows | Adding codeball
+- (workflow) #fse-512 | github/workflows | Adding codeball
 - (tests) #fse-509 | apps/assets 1.0.2 apps/governance 1.0.2 apps/staking 1.0.2 packages/evmos-wallet 1.0.4 packages/helpers 1.0.2 packages/services 1.0.1 packages/ui-helpers 1.0.1 | Adding shared package for jest related configuration
 - (design) #229 | apps/staking 1.0.1 apps/assets 1.0.1 | Making design fixes so we don't have extra scrollbars
 - (improvements) #fse-487 | evmos-wallet 1.0.3 | Use preferNoSetFee while signing with Keplr and chain is Evmos